+++ test/torchaudio_unittest/kaldi_compatibility_cpu_test.py
@@ -1,7 +1,7 @@
 import torch
 
 from torchaudio_unittest import common_utils
-from .kaldi_compatibility_impl import Kaldi, KaldiCPUOnly
+from .kaldi_compatibility_impl import Kaldi
 
 
 class TestKaldiFloat32(Kaldi, common_utils.PytorchTestCase):
@@ -11,9 +11,4 @@ class TestKaldiFloat32(Kaldi, common_utils.PytorchTestCase):
 
 class TestKaldiFloat64(Kaldi, common_utils.PytorchTestCase):
     dtype = torch.float64
-    device = torch.device('cpu')
-
-
-class TestKaldiCPUOnly(KaldiCPUOnly, common_utils.PytorchTestCase):
-    dtype = torch.float32
     device = torch.device('cpu')
